+++ firebase-installations/src/main/java/com/google/firebase/installations/remote/FirebaseInstallationServiceClient.java
@@ -30,6 +30,7 @@ import com.google.android.gms.common.util.Hex;
 import com.google.android.gms.common.util.VisibleForTesting;
 import com.google.firebase.heartbeatinfo.HeartBeatInfo;
 import com.google.firebase.heartbeatinfo.HeartBeatInfo.HeartBeat;
+import com.google.firebase.inject.Provider;
 import com.google.firebase.installations.FirebaseInstallationsException;
 import com.google.firebase.installations.FirebaseInstallationsException.Status;
 import com.google.firebase.installations.remote.InstallationResponse.ResponseCode;
@@ -101,13 +102,13 @@ public class FirebaseInstallationServiceClient {
   static final String PARSING_EXPIRATION_TIME_ERROR_MESSAGE = "Invalid Expiration Timestamp.";
 
   private final Context context;
-  private final UserAgentPublisher userAgentPublisher;
-  private final HeartBeatInfo heartbeatInfo;
+  private final Provider<UserAgentPublisher> userAgentPublisher;
+  private final Provider<HeartBeatInfo> heartbeatInfo;
 
   public FirebaseInstallationServiceClient(
       @NonNull Context context,
-      @Nullable UserAgentPublisher publisher,
-      @Nullable HeartBeatInfo heartbeatInfo) {
+      @NonNull Provider<UserAgentPublisher> publisher,
+      @NonNull Provider<HeartBeatInfo> heartbeatInfo) {
     this.context = context;
     this.userAgentPublisher = publisher;
     this.heartbeatInfo = heartbeatInfo;
@@ -173,7 +174,7 @@ public class FirebaseInstallationServiceClient {
 
         // Return empty installation response with BAD_CONFIG response code after max retries
         return InstallationResponse.builder().setResponseCode(ResponseCode.BAD_CONFIG).build();
-      } catch (IOException ignored) {
+      } catch (AssertionError | IOException ignored) {
         retryCount++;
       } finally {
         httpURLConnection.disconnect();
@@ -394,7 +395,8 @@ public class FirebaseInstallationServiceClient {
         logBadConfigError();
 
         return TokenResult.builder().setResponseCode(TokenResult.ResponseCode.BAD_CONFIG).build();
-      } catch (IOException ignored) {
+        // TODO(b/166168291): Remove code duplication and clean up this class.
+      } catch (AssertionError | IOException ignored) {
         retryCount++;
       } finally {
         httpURLConnection.disconnect();
@@ -432,10 +434,12 @@ public class FirebaseInstallationServiceClient {
     httpURLConnection.addRequestProperty(CONTENT_ENCODING_HEADER_KEY, GZIP_CONTENT_ENCODING);
     httpURLConnection.addRequestProperty(CACHE_CONTROL_HEADER_KEY, CACHE_CONTROL_DIRECTIVE);
     httpURLConnection.addRequestProperty(X_ANDROID_PACKAGE_HEADER_KEY, context.getPackageName());
-    if (heartbeatInfo != null && userAgentPublisher != null) {
-      HeartBeat heartbeat = heartbeatInfo.getHeartBeatCode(FIREBASE_INSTALLATIONS_ID_HEARTBEAT_TAG);
+    if ((heartbeatInfo.get() != null) && (userAgentPublisher.get() != null)) {
+      HeartBeat heartbeat =
+          heartbeatInfo.get().getHeartBeatCode(FIREBASE_INSTALLATIONS_ID_HEARTBEAT_TAG);
       if (heartbeat != HeartBeat.NONE) {
-        httpURLConnection.addRequestProperty(USER_AGENT_HEADER, userAgentPublisher.getUserAgent());
+        httpURLConnection.addRequestProperty(
+            USER_AGENT_HEADER, userAgentPublisher.get().getUserAgent());
         httpURLConnection.addRequestProperty(
             HEART_BEAT_HEADER, Integer.toString(heartbeat.getCode()));
       }
@@ -446,11 +450,14 @@ public class FirebaseInstallationServiceClient {
   }
 
   // Read the response from the createFirebaseInstallation API.
-  private InstallationResponse readCreateResponse(HttpURLConnection conn) throws IOException {
+  private InstallationResponse readCreateResponse(HttpURLConnection conn)
+      throws AssertionError, IOException {
     InputStream inputStream = conn.getInputStream();
     JsonReader reader = new JsonReader(new InputStreamReader(inputStream, UTF_8));
     TokenResult.Builder tokenResult = TokenResult.builder();
     InstallationResponse.Builder builder = InstallationResponse.builder();
+    // JsonReader.peek will sometimes throw AssertionErrors in Android 8.0 and above. See
+    // https://b.corp.google.com/issues/79920590 for details.
     reader.beginObject();
     while (reader.hasNext()) {
       String name = reader.nextName();
@@ -486,10 +493,13 @@ public class FirebaseInstallationServiceClient {
   }
 
   // Read the response from the generateAuthToken FirebaseInstallation API.
-  private TokenResult readGenerateAuthTokenResponse(HttpURLConnection conn) throws IOException {
+  private TokenResult readGenerateAuthTokenResponse(HttpURLConnection conn)
+      throws AssertionError, IOException {
     InputStream inputStream = conn.getInputStream();
     JsonReader reader = new JsonReader(new InputStreamReader(inputStream, UTF_8));
     TokenResult.Builder builder = TokenResult.builder();
+    // JsonReader.peek will sometimes throw AssertionErrors in Android 8.0 and above. See
+    // https://b.corp.google.com/issues/79920590 for details.
     reader.beginObject();
     while (reader.hasNext()) {
       String name = reader.nextName();